+++ src/balances/balances.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { BalancesService } from './balances.service';
 import { BalancesController } from './balances.controller';
-import { TransactionServiceModule } from '../services/transaction-service/transaction-service.module';
-import { ConfigServiceModule } from '../services/config-service/config-service.module';
-import { ExchangeModule } from '../services/exchange/exchange.module';
+import { TransactionServiceModule } from '../datasources/transaction-service/transaction-service.module';
+import { ConfigServiceModule } from '../datasources/config-service/config-service.module';
+import { ExchangeModule } from '../datasources/exchange/exchange.module';
 
 @Module({
   controllers: [BalancesController],
