+++ src/chains/chains.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ChainsService } from './chains.service';
 import { ChainsController } from './chains.controller';
-import { ConfigServiceModule } from '../services/config-service/config-service.module';
-import { TransactionServiceModule } from '../services/transaction-service/transaction-service.module';
+import { ConfigServiceModule } from '../datasources/config-service/config-service.module';
+import { TransactionServiceModule } from '../datasources/transaction-service/transaction-service.module';
 
 @Module({
   imports: [ConfigServiceModule, TransactionServiceModule],
