+++ src/datasources/config-service/entities/native.currency.entity.ts
